+++ lib/item_repository.rb
@@ -1,7 +1,11 @@
 class ItemRepository
 
   def initialize(items)
-    @items = items
+    @items = populate_items_array
+  end
+
+  def populate_items_array(items)
+    items.map { |item| Item.new(item) }
   end
 
   def all
@@ -36,4 +40,3 @@ class ItemRepository
     "#<#{self.class} #{@items.size} rows>"
   end
 end
-
